+++ oteapi/models/__init__.py
@@ -8,10 +8,11 @@ from .datacacheconfig import DataCacheConfig
 from .filterconfig import FilterConfig
 from .functionconfig import FunctionConfig
 from .genericconfig import AttrDict, GenericConfig
-from .mappingconfig import MappingConfig
+from .mappingconfig import MappingConfig, RDFTriple
 from .resourceconfig import ResourceConfig
 from .sessionupdate import SessionUpdate
 from .transformationconfig import TransformationConfig, TransformationStatus
+from .triplestoreconfig import TripleStoreConfig
 
 __all__ = (
     "AttrDict",
@@ -20,11 +21,13 @@ __all__ = (
     "FunctionConfig",
     "GenericConfig",
     "MappingConfig",
+    "RDFTriple",
     "ResourceConfig",
     "SessionUpdate",
     "StrategyConfig",
     "TransformationConfig",
     "TransformationStatus",
+    "TripleStoreConfig",
 )
 
 StrategyConfig = Union[