+++ oteapi/models/mappingconfig.py
@@ -1,27 +1,28 @@
 """Pydantic Mapping Configuration Data Model."""
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, Tuple
 
-from pydantic import Field, conlist
+from pydantic import Field
 
 from oteapi.models.genericconfig import GenericConfig
 
-SemanticTriple = conlist(str, min_items=3, max_items=3)
+RDFTriple = Tuple[str, str, str]
 
 
 class MappingConfig(GenericConfig):
     """Mapping Strategy Data Configuration."""
 
     mappingType: str = Field(
-        ..., description="Type of registered mapping strategy. E.g., `mapping/demo`."
+        ...,
+        description="Type of registered mapping strategy.",
     )
     prefixes: Optional[Dict[str, str]] = Field(
         None,
         description=(
-            "List of shortnames that expands to an IRI "
-            "given as local value/IRI-expansion-pairs."
+            "List of shortnames that expands to an IRI given as local "
+            "value/IRI-expansion-pairs."
         ),
     )
-    triples: Optional[List[SemanticTriple]] = Field(  # type: ignore[valid-type]
+    triples: Optional[List[RDFTriple]] = Field(
         None,
-        description="List of semantic triples given as (subject, predicate, object).",
+        description="List of RDF triples given as (subject, predicate, object).",
     )