+++ oteapi/strategies/filter/crop_filter.py
@@ -14,12 +14,19 @@ if TYPE_CHECKING:  # pragma: no cover
 class CropImageConfig(AttrDict):
     """Configuration model for crop data."""
 
-    crop: Tuple[int, int, int, int] = Field(..., description="Box cropping parameters.")
+    crop: Tuple[int, int, int, int] = Field(
+        ..., description="Box cropping parameters (left, top, right, bottom)."
+    )
 
 
 class CropImageFilterConfig(FilterConfig):
     """Crop filter strategy filter config."""
 
+    filterType: str = Field(
+        "filter/crop",
+        const=True,
+        description=FilterConfig.__fields__["filterType"].field_info.description,
+    )
     configuration: CropImageConfig = Field(
         ..., description="Image crop filter strategy-specific configuration."
     )
@@ -29,7 +36,7 @@ class SessionUpdateCropFilter(SessionUpdate):
     """Return model for `CropImageFilter`."""
 
     imagecrop: Tuple[int, int, int, int] = Field(
-        ..., description="Box cropping parameters."
+        ..., description="Box cropping parameters (left, top, right, bottom)."
     )
 
 
@@ -45,12 +52,18 @@ class CropImageFilter:
 
     filter_config: CropImageFilterConfig
 
-    def initialize(self, session: "Optional[Dict[str, Any]]" = None) -> SessionUpdate:
+    def initialize(
+        self,
+        session: "Optional[Dict[str, Any]]" = None,
+    ) -> SessionUpdateCropFilter:
         """Initialize strategy and return a dictionary."""
-        return SessionUpdate()
+        return SessionUpdateCropFilter(
+            imagecrop=self.filter_config.configuration.crop,
+        )
 
     def get(
-        self, session: "Optional[Dict[str, Any]]" = None
-    ) -> SessionUpdateCropFilter:
+        self,
+        session: "Optional[Dict[str, Any]]" = None,
+    ) -> SessionUpdate:
         """Execute strategy and return a dictionary"""
-        return SessionUpdateCropFilter(imagecrop=self.filter_config.configuration.crop)
+        return SessionUpdate()