+++ oteapi/strategies/filter/sql_query_filter.py
@@ -14,6 +14,11 @@ if TYPE_CHECKING:  # pragma: no cover
 class SqlQueryFilterConfig(FilterConfig):
     """SQLite query filter strategy filter config."""
 
+    filterType: str = Field(
+        "filter/sql",
+        const=True,
+        description=FilterConfig.__fields__["filterType"].field_info.description,
+    )
     query: str = Field(..., description="A SQL query string.")
 
 
@@ -35,10 +40,13 @@ class SQLQueryFilter:
 
     filter_config: SqlQueryFilterConfig
 
-    def initialize(self, session: "Optional[Dict[str, Any]]" = None) -> SessionUpdate:
+    def initialize(
+        self,
+        session: "Optional[Dict[str, Any]]" = None,
+    ) -> SessionUpdateSqlQuery:
         """Initialize strategy."""
-        return SessionUpdate()
-
-    def get(self, session: "Optional[Dict[str, Any]]" = None) -> SessionUpdateSqlQuery:
-        """Execute strategy and return a dictionary."""
         return SessionUpdateSqlQuery(sqlquery=self.filter_config.query)
+
+    def get(self, session: "Optional[Dict[str, Any]]" = None) -> SessionUpdate:
+        """Execute strategy and return a dictionary."""
+        return SessionUpdate()